Refactor + add pre-caching
